+++ custom_components/mydolphin_plus/binary_sensor.py
@@ -1,5 +1,4 @@
 import logging
-import sys
 
 from homeassistant.components.binary_sensor import (
     BinarySensorEntity,
@@ -45,31 +44,18 @@ class MyDolphinPlusBinarySensorEntity(MyDolphinPlusBaseEntity, BinarySensorEntit
 
         self._attr_device_class = entity_description.device_class
 
-    def _handle_coordinator_update(self) -> None:
+    def update_component(self, data):
         """Fetch new state parameters for the sensor."""
-        try:
-            device_data = self.get_data()
+        if data is not None:
+            is_on = data.get(ATTR_IS_ON)
+            attributes = data.get(ATTR_ATTRIBUTES)
+            icon = data.get(ATTR_ICON)
 
-            if device_data is not None:
-                is_on = device_data.get(ATTR_IS_ON)
-                attributes = device_data.get(ATTR_ATTRIBUTES)
-                icon = device_data.get(ATTR_ICON)
+            self._attr_is_on = is_on
+            self._attr_extra_state_attributes = attributes
 
-                self._attr_is_on = is_on
-                self._attr_extra_state_attributes = attributes
+            if icon is not None:
+                self._attr_icon = icon
 
-                if icon is not None:
-                    self._attr_icon = icon
-
-            else:
-                self._attr_is_on = None
-
-            self.async_write_ha_state()
-
-        except Exception as ex:
-            exc_type, exc_obj, tb = sys.exc_info()
-            line_number = tb.tb_lineno
-
-            _LOGGER.error(
-                f"Failed to update {self.unique_id}, Error: {ex}, Line: {line_number}"
-            )
+        else:
+            self._attr_is_on = None