+++ custom_components/mydolphin_plus/light.py
@@ -1,6 +1,5 @@
 from abc import ABC
 import logging
-import sys
 from typing import Any
 
 from homeassistant.components.light import LightEntity, LightEntityDescription
@@ -51,27 +50,14 @@ class MyDolphinPlusLightEntity(MyDolphinPlusBaseEntity, LightEntity, ABC):
     async def async_turn_off(self, **kwargs: Any) -> None:
         await self.async_execute_device_action(ACTION_ENTITY_TURN_OFF)
 
-    def _handle_coordinator_update(self) -> None:
+    def update_component(self, data):
         """Fetch new state parameters for the sensor."""
-        try:
-            device_data = self.get_data()
+        if data is not None:
+            is_on = data.get(ATTR_IS_ON)
+            attributes = data.get(ATTR_ATTRIBUTES)
 
-            if device_data is not None:
-                is_on = device_data.get(ATTR_IS_ON)
-                attributes = device_data.get(ATTR_ATTRIBUTES)
+            self._attr_is_on = is_on
+            self._attr_extra_state_attributes = attributes
 
-                self._attr_is_on = is_on
-                self._attr_extra_state_attributes = attributes
-
-            else:
-                self._attr_is_on = None
-
-            self.async_write_ha_state()
-
-        except Exception as ex:
-            exc_type, exc_obj, tb = sys.exc_info()
-            line_number = tb.tb_lineno
-
-            _LOGGER.error(
-                f"Failed to update {self.unique_id}, Error: {ex}, Line: {line_number}"
-            )
+        else:
+            self._attr_is_on = None