+++ custom_components/mydolphin_plus/number.py
@@ -1,6 +1,5 @@
 from abc import ABC
 import logging
-import sys
 
 from homeassistant.components.number import NumberEntity, NumberEntityDescription
 from homeassistant.config_entries import ConfigEntry
@@ -46,27 +45,14 @@ class MyDolphinPlusNumberEntity(MyDolphinPlusBaseEntity, NumberEntity, ABC):
         """Change the selected option."""
         await self.async_execute_device_action(ACTION_ENTITY_SET_NATIVE_VALUE, value)
 
-    def _handle_coordinator_update(self) -> None:
+    def update_component(self, data):
         """Fetch new state parameters for the sensor."""
-        try:
-            device_data = self.get_data()
+        if data is not None:
+            state = data.get(ATTR_STATE)
+            attributes = data.get(ATTR_ATTRIBUTES)
 
-            if device_data is not None:
-                state = device_data.get(ATTR_STATE)
-                attributes = device_data.get(ATTR_ATTRIBUTES)
+            self._attr_native_value = state
+            self._attr_extra_state_attributes = attributes
 
-                self._attr_native_value = state
-                self._attr_extra_state_attributes = attributes
-
-            else:
-                self._attr_native_value = None
-
-            self.async_write_ha_state()
-
-        except Exception as ex:
-            exc_type, exc_obj, tb = sys.exc_info()
-            line_number = tb.tb_lineno
-
-            _LOGGER.error(
-                f"Failed to update {self.unique_id}, Error: {ex}, Line: {line_number}"
-            )
+        else:
+            self._attr_native_value = None